Deploy to Docker Swarm with separate `.yaml` file
Stack can be run in Docker Swarm mode using a `.yaml` file very similar
to that used by Docker Compose. Perhaps the most prevalent difference is
that Swarm does not support building images as Compose does, so each
service must be specified with a pre-build repository image.

The Swarm equivalent to `docker-compose up` is `docker stack deploy`,
which requires both a `--compose-file FILE` option and a `STACKNAME`
argument.

For some reason the stack often takes exceedingly long to spin up. The
`web` service is usually accessible within a few seconds of `deploy`ment
but usually returns a `502 BAD GATEWAY` code indicating that the `api`
service is unresponsive. Inspecting the `api` service logs (`docker
service logs groceries_api`) confirms that the `entrypoint.sh` script
seems to be waiting for the `db` service to start, though the latter's
logs indicate it is running and listening on the expected port as
quickly as the `web` service. After as long as several minutes, the
`api` service will finally reach the `db` service and complete its own
startup, at which point the `web` service will successfully serve the
frontend to a browser.

The delay is perplexing and its cause is unclear. When run using
`docker-compose`, all services are responsive almost immediately. It is
noted that, while troubleshooting this issue, there were two or three
trials in which the stack *did* deploy as rapidly as expected, though
why these deployments were any different from others is equally as
unclear.

It may be relevant that Swarm uses a different network driver (overlay)
than Compose (bridge) to fulfill its purpose of coordinating multiple
hosts, though initial review of the available documentation hasn't
revealed an obvious explanation or solution.