Migrate credentials to secrets in Docker Swarm mode
Most of this implementation is docker secrets "by the books". Separate
`.yaml` files are maintained for Compose and Stack despite the
divergence and redundancy -- means of merging these two files will be
explored.

A few things to note: 1) database initialization is not done via shell
script as there is no well-documented way of reading from file in an sql
script. The solution here seems a bit hack-y and may break with changes
to `createuser`, but this was the best solution that could be found in
reasoable time. 2) The `djrf` system user (not the Postgres role) is
created with a specified uid/gid in `api/Dockerfile` since this id is
also used to set permissions on files in `/run/secrets/`. 3) Django
reads credentials from file in `settings.py` instead of in
`server_secrets.py` since the file locations are well-known via public
Docker documentation (spoiler: they are derived from the names given to
secrets in `.yaml` files).